requirements.txt
numpy>=1.23
opencv-python>=4.8
matplotlib>=3.7
tqdm>=4.66
pyyaml>=6.0
pandas